[chore] renamed files according to app convention
