+++ src/components/Footer.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import moment from 'moment'
 import { Link } from 'react-router-dom'
-import Socials from './Socials'
+import Socials from './Socials.jsx'
 
 export const Footer = ({ netId }) => {
   const footerClassName = netId === '77' ? 'sokol' : ''
