+++ src/components/Header.jsx
@@ -5,8 +5,8 @@ import menuIconBase from '../assets/images/icons/icon-menu.svg'
 import menuIconSokol from '../assets/images/icons/icon-menu-sokol.svg'
 import menuOpenIconBase from '../assets/images/icons/icon-close.svg'
 import menuOpenIconSokol from '../assets/images/icons/icon-close-sokol.svg'
-import NavigationLinks from './NavigationLinks'
-import MobileMenuLinks from './MobileMenuLinks'
+import NavigationLinks from './NavigationLinks.jsx'
+import MobileMenuLinks from './MobileMenuLinks.jsx'
 
 export const Header = ({ netId, baseRootPath, navigationData, showMobileMenu, onMenuToggle }) => {
   const headerClassName = netId === '77' ? 'sokol' : ''
