+++ src/components/NavigationLinks.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { NavLink } from 'react-router-dom'
 
-const NavigationLinks = ({ navigationData, baseRootPath }) => {
+const NavigationLinks = ({ navigationData }) => {
   return navigationData.map((item, index) => (
     <NavLink activeClassName="active" className="link" exact key={index} to={item.url}>
       <i className={`link-icon ${item.icon}`} />
